pyproject.toml
[tool.poetry]
name = "minechat-connector"
version = "0.1.0"
description = ""
authors = ["Elnar Mengelbaev <elnar_men@mail.ru>"]
readme = "README.md"
packages = [{include = "minechat_connector"}]

[tool.poetry.dependencies]
python = "^3.10"
aiofiles = '23.2.1'
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
